Move interface transceiver presence to interface_Transceiver_cli.go
