--- show_client/interface_transceiver_presence.go
@@ -1,70 +1,0 @@
-package show_client
-
-import (
-	"encoding/json"
-
-	log "github.com/golang/glog"
-	sdc "github.com/sonic-net/sonic-gnmi/sonic_data_client"
-)
-
-func getAllPortsFromConfigDB() ([]string, error) {
-	queries := [][]string{
-		{"CONFIG_DB", "PORT"},
-	}
-	data, err := GetMapFromQueries(queries)
-	if err != nil {
-		log.Errorf("Unable to get data from CONFIG_DB queries %v, got err: %v", queries, err)
-		return nil, err
-	}
-	log.V(6).Infof("Data from CONFIG_DB: %v", data)
-
-	ports := make([]string, 0, len(data))
-	for iface := range data {
-		ports = append(ports, iface)
-	}
-	return ports, nil
-}
-
-func getInterfaceTransceiverPresence(options sdc.OptionMap) ([]byte, error) {
-	intf, _ := options["interface"].String()
-
-	// Get STATE_DB transceiver info
-	queries := [][]string{
-		{"STATE_DB", "TRANSCEIVER_INFO"},
-	}
-	data, err := GetMapFromQueries(queries)
-	if err != nil {
-		log.Errorf("Unable to get transceiver data from STATE_DB queries %v, got err: %v", queries, err)
-		return nil, err
-	}
-	log.V(6).Infof("TRANSCEIVER_INFO Data from STATE_DB: %v", data)
-
-	status := make(map[string]string)
-
-	if intf != "" {
-		// If specific interface provided, skip ConfigDB check
-		if _, exist := data[intf]; exist {
-			status[intf] = "Present"
-		} else {
-			status[intf] = "Not Present"
-		}
-	} else {
-		// No specific interface provided, get all from ConfigDB
-		ports, err := getAllPortsFromConfigDB()
-		if err != nil {
-			log.Errorf("Unable to get all ports from CONFIG_DB, %v", err)
-			return nil, err
-		}
-
-		for _, port := range ports {
-			if _, exist := data[port]; exist {
-				status[port] = "Present"
-			} else {
-				status[port] = "Not Present"
-			}
-		}
-	}
-
-	log.V(6).Infof("Transceiver presence status: %v", status)
-	return json.Marshal(status)
-}
